refactor: 투표 생성 리팩토링
