+++ domain/src/main/java/com/programmers/bucketback/domains/vote/model/VoteCreateImplRequest.java
@@ -1,0 +1,11 @@
+package com.programmers.bucketback.domains.vote.model;
+
+import com.programmers.bucketback.Hobby;
+
+public record VoteCreateImplRequest(
+	Hobby hobby,
+	String content,
+	Long item1Id,
+	Long item2Id
+) {
+}
